Fix GridLayer occassionally leaving tiles from two layers active
If GridLayer finds a tile that is not current but is still active,
set a flag to do a prune if at the end of the fade.

I suspect the original problem might be due to a race between
_updateOpacity and timed prune from _tileReady. As it seems cheap
to calculate when a clearing prune (maybe) needed at the end of
the fade, just do it.